Introduce `DriveOperations` class that holds all drive operations.
